"Communicated with client on New web services and made some changes according to client feedback. Implemented SOPCategory and SOPDetail services. Removed Registration, Guest Name, Certificate Name dropdown property, Added Send Text message in Update user profile Added Crittercism library"

+++ Code/GoBoardPro/View/Graph/GraphDetailViewController.m
@@ -189,7 +189,7 @@
             NSDictionary *aDict = [mutArrDataSource objectAtIndex:i];
             [(UILabel *)[lbl objectAtIndex:i] setText:[[aDict objectForKey:@"xValue"] objectAtIndex:indexPath.row]];
         }
-        [aLblTotalValue setText:[NSString stringWithFormat:@"%d", [aLblFistValue.text integerValue] + [aLblSecondValue.text integerValue] + [aLblThirdValue.text integerValue]]];
+        [aLblTotalValue setText:[NSString stringWithFormat:@"%ld", (long)([aLblFistValue.text integerValue] + [aLblSecondValue.text integerValue] + [aLblThirdValue.text integerValue])]];
     }
    
     return aCell;
